requirements.txt
# This is a requirements.txt file for Python's pip
# You can run 'pip install -r requirements.txt' to install all requirements.
flask==0.10.1
py-bcrypt==0.4